BigDecimal implementation

+++ src/main/kotlin/com/jakubspiewak/robotarm/robot/RobotArmLengths.kt
@@ -1,5 +1,7 @@
 package com.jakubspiewak.robotarm.robot
 
+import java.math.BigDecimal
+
 /*
 elements: ║ ═
 
@@ -16,15 +18,15 @@ elements: ║ ═
  */
 data class RobotArmLengths(
     // length between join j0 and join j1
-    val l0: Double,
+    val l0: BigDecimal,
     // length between join j1 and join j2
-    val l1: Double,
+    val l1: BigDecimal,
     // length between join j2 and join j3
-    val l2: Double,
+    val l2: BigDecimal,
     // length between join j3 and join j4
-    val l3: Double,
+    val l3: BigDecimal,
     // length between join j4and join j5
-    val l4: Double,
+    val l4: BigDecimal,
     // length between join j5 and end-effector
-    val l5: Double
+    val l5: BigDecimal
 )
